app/requirements.txt
semantic-kernel
pyautogen==0.2.0b1
azure-cosmos
azure-identity
openai>=1.0.0b3